requirements.txt
python-telegram-bot==11.1.
redis==3.5.3
vk-api==11.9.0